backend/pyproject.toml
[tool.poetry]
name = "rl-demo-web"
version = "0.1.0"
description = ""
authors = ["momonong <morrischen2.9@gmail.com>"]
readme = "README.md"
packages = [{include = "rl_demo_web"}]

[tool.poetry.dependencies]
python = "3.10.*"
pillow = "^10.0.0"
pandas = "^2.0.3"
fastapi = {version = "0.99.1", extras = ["all"]}
scikit-learn = "^1.3.0"
matplotlib = "^3.7.2"
tensorflow = "2.8.0"
protobuf = "3.20.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
